// src/app/auth/interfaces/login.interface.ts
// Generated by https://quicktype.io

export interface InterfazLogin {
    ok?:boolean;
    uid?: string;
    name?: string;
    email?: string;
    token?: string;
    msg?: string;
    idRole: number;
}

// Generated by https://quicktype.io

export interface Login {
    resutls: Resutl[];
}

export interface Resutl {
    idLog:       number;
    description: string;
    user_name:   string;
    date:        string;
}